ingest/requirements.txt
pydantic
langchain>=0.2.10
langchain-core>=0.2.10
langchain-community>=0.2.10
sentence-transformers==2.2.2
huggingface-hub==0.10.1
faiss-cpu>=1.8.0
numpy>=1.26
tqdm>=4.66.0
pypdf>=4.2.0
python-docx>=1.0.1
python-pptx>=0.6.23